+++ src/collective/eeafaceted/collectionwidget/browser/views.py
@@ -8,7 +8,8 @@ class RenderCategoryView(BrowserView):
     def __init__(self, context, request):
         ''' '''
         BrowserView.__init__(self, context, request)
-        self.portal_url = getToolByName(self.context, 'portal_url').getPortalObject().absolute_url()
+        self.portal = getToolByName(self.context, 'portal_url').getPortalObject()
+        self.portal_url = self.portal.absolute_url()
 
     def __call__(self, widget):
         self.widget = widget